[Refactor] apply custom hook #775
