+++ components/Layout/Layout.tsx
@@ -1,13 +1,8 @@
 import Link from 'next/link';
 import { useRouter } from 'next/router';
-import { useEffect } from 'react';
-import { useRecoilState, useSetRecoilState } from 'recoil';
-import { userState, liveState } from 'utils/recoil/layout';
-import { reloadMatchState, openCurrentMatchState } from 'utils/recoil/match';
-import { errorState } from 'utils/recoil/error';
-import { modalState } from 'utils/recoil/modal';
-import { seasonListState } from 'utils/recoil/seasons';
-import instance from 'utils/axios';
+import { useRecoilValue } from 'recoil';
+import { userState } from 'utils/recoil/layout';
+import { openCurrentMatchState } from 'utils/recoil/match';
 import Statistics from 'pages/statistics';
 import Header from './Header';
 import Footer from './Footer';
@@ -17,93 +12,25 @@ import AdminReject from '../admin/AdminReject';
 import styles from 'styles/Layout/Layout.module.scss';
 import useAxiosWithToast from 'hooks/useAxiosWithToast';
 
+import useAnnouncementCheck from 'hooks/Layout/useAnnouncementCheck';
+import useSetAfterGameModal from 'hooks/Layout/useSetAfterGameModal';
+import useGetUserInfo from 'hooks/Layout/useGetUserInfo';
+import useLiveCheck from 'hooks/Layout/useLiveCheck';
+
 type AppLayoutProps = {
   children: React.ReactNode;
 };
 
 export default function AppLayout({ children }: AppLayoutProps) {
-  const [user, setUser] = useRecoilState(userState);
-  const [live, setLive] = useRecoilState(liveState);
-  const [openCurrentMatch, setOpenCurrentMatch] = useRecoilState(
-    openCurrentMatchState
-  );
-  const [reloadMatch, setReloadMatch] = useRecoilState(reloadMatchState);
-  const setSeasonList = useSetRecoilState(seasonListState);
-  const setError = useSetRecoilState(errorState);
-  const setModal = useSetRecoilState(modalState);
+  const user = useRecoilValue(userState);
+  const openCurrentMatch = useRecoilValue(openCurrentMatchState);
   const presentPath = useRouter().asPath;
-  const announcementTime = localStorage.getItem('announcementTime');
 
   useAxiosWithToast();
-
-  useEffect(() => {
-    getUserHandler();
-    getSeasonListHandler();
-  }, []);
-
-  const getAnnouncementHandler = async () => {
-    try {
-      const res = await instance.get(`/pingpong/announcement`);
-      res.data.content !== '' &&
-        setModal({
-          modalName: 'EVENT-ANNOUNCEMENT',
-          announcement: res.data,
-        });
-    } catch (e) {
-      setError('RJ01');
-    }
-  };
-  const getSeasonListHandler = async () => {
-    try {
-      const res = await instance.get(`/pingpong/seasonlist`);
-      setSeasonList({ ...res?.data });
-    } catch (e) {
-      setError('DK02');
-    }
-  };
-
-  useEffect(() => {
-    if (presentPath === '/') {
-      if (
-        !announcementTime ||
-        Date.parse(announcementTime) < Date.parse(new Date().toString())
-      )
-        getAnnouncementHandler();
-    } else setModal({ modalName: null });
-  }, [presentPath]);
-
-  useEffect(() => {
-    if (user.intraId) {
-      getLiveHandler();
-      if (reloadMatch) setReloadMatch(false);
-    }
-  }, [presentPath, user, reloadMatch]);
-
-  useEffect(() => {
-    if (live?.event === 'match') setOpenCurrentMatch(true);
-    else {
-      if (live?.event === 'game') setModal({ modalName: 'FIXED-AFTER_GAME' });
-      setOpenCurrentMatch(false);
-    }
-  }, [live]);
-
-  const getUserHandler = async () => {
-    try {
-      const res = await instance.get(`/pingpong/users`);
-      setUser(res?.data);
-    } catch (e) {
-      setError('JB02');
-    }
-  };
-
-  const getLiveHandler = async () => {
-    try {
-      const res = await instance.get(`/pingpong/users/live`);
-      setLive({ ...res?.data });
-    } catch (e) {
-      setError('JB03');
-    }
-  };
+  useGetUserInfo();
+  useSetAfterGameModal();
+  // useLiveCheck(presentPath);
+  useAnnouncementCheck(presentPath);
 
   return presentPath.includes('/admin') ? (
     user.isAdmin ? (
