+++ components/main/SearchBar.tsx
@@ -1,58 +1,20 @@
 import Link from 'next/link';
-import { useEffect, useState, useRef } from 'react';
-import { useSetRecoilState } from 'recoil';
-import instance from 'utils/axios';
-import { errorState } from 'utils/recoil/error';
 import { GoSearch } from 'react-icons/go';
 import { IoIosCloseCircle } from 'react-icons/io';
 import styles from 'styles/main/SearchBar.module.scss';
 
-let timer: ReturnType<typeof setTimeout>;
+import useSearchBar from 'hooks/useSearchBar';
 
 export default function SearchBar() {
-  const [keyword, setKeyword] = useState<string>('');
-  const [showDropDown, setShowDropDown] = useState<boolean>(false);
-  const [searchResult, setSearchResult] = useState<string[]>([]);
-  const setError = useSetRecoilState(errorState);
-  const searchBarRef = useRef<HTMLDivElement>(null);
-
-  useEffect(() => {
-    const checkId = /^[a-z|A-Z|0-9|-]+$/;
-    if (keyword === '' || (keyword.length && !checkId.test(keyword))) {
-      clearTimeout(timer);
-      setSearchResult([]);
-    } else if (checkId.test(keyword)) {
-      debounce(getSearchResultHandler, 500)();
-    }
-  }, [keyword]);
-
-  useEffect(() => {
-    document.addEventListener('mousedown', handleClickOutside);
-    return () => {
-      document.removeEventListener('mousedown', handleClickOutside);
-    };
-  }, []);
-
-  const getSearchResultHandler = async () => {
-    try {
-      const res = await instance.get(`/pingpong/users/searches?q=${keyword}`);
-      setSearchResult(res?.data.users);
-    } catch (e) {
-      setError('JB06');
-    }
-  };
-
-  const handleClickOutside = (event: MouseEvent) => {
-    if (
-      searchBarRef.current &&
-      !searchBarRef.current.contains(event.target as Node)
-    )
-      setShowDropDown(false);
-  };
-
-  const keywordHandler = (event: React.ChangeEvent<HTMLInputElement>) => {
-    setKeyword(event.target.value);
-  };
+  const [
+    keyword,
+    setKeyword,
+    keywordHandler,
+    showDropDown,
+    setShowDropDown,
+    searchResult,
+    searchBarRef,
+  ] = useSearchBar();
 
   return (
     <div className={styles.searchBar} ref={searchBarRef}>
@@ -90,13 +52,4 @@ export default function SearchBar() {
       )}
     </div>
   );
-}
-
-function debounce(callback: () => void, timeout: number) {
-  return () => {
-    clearTimeout(timer);
-    timer = setTimeout(() => {
-      callback();
-    }, timeout);
-  };
 }
