+++ components/modal/menu/LogoutModal.tsx
@@ -4,20 +4,10 @@ import { loginState } from 'utils/recoil/login';
 import { modalState } from 'utils/recoil/modal';
 import styles from 'styles/modal/menu/LogoutModal.module.scss';
 
+import useLogoutCheck from 'hooks/Login/useLogoutCheck';
+
 export default function LogoutModal() {
-  const setIsLoggedIn = useSetRecoilState(loginState);
-  const setModal = useSetRecoilState(modalState);
-  const router = useRouter();
-
-  const onReturn = () => {
-    setModal({ modalName: null });
-  };
-
-  const onLogout = () => {
-    localStorage.removeItem('42gg-token');
-    setIsLoggedIn(false);
-    router.push(`/`);
-  };
+  const [onReturn, onLogout] = useLogoutCheck();
 
   return (
     <div className={styles.container}>
