+++ components/rank/RankList.tsx
@@ -1,14 +1,13 @@
-import { useEffect, useState } from 'react';
-import { useRecoilState, useRecoilValue, useSetRecoilState } from 'recoil';
+import { useState } from 'react';
+import { useRecoilValue } from 'recoil';
 import { MatchMode } from 'types/mainType';
 import { RankUser, NormalUser, Rank } from 'types/rankTypes';
-import { myRankState, scrollState } from 'utils/recoil/myRank';
 import { seasonListState } from 'utils/recoil/seasons';
-import { errorState } from 'utils/recoil/error';
-import instance from 'utils/axios';
 import RankListMain from './topRank/RankListMain';
 import RankListFrame from './RankListFrame';
 import RankListItem from './RankListItem';
+
+import useRankList from 'hooks/rank/useRankList';
 
 interface RankListProps {
   toggleMode: MatchMode;
@@ -20,10 +19,7 @@ export default function RankList({
   season,
   isMain = false,
 }: RankListProps) {
-  const [myRank, setMyRank] = useRecoilState(myRankState);
-  const [isScroll, setIsScroll] = useRecoilState(scrollState);
   const { seasonMode } = useRecoilValue(seasonListState);
-  const setError = useSetRecoilState(errorState);
   const [rank, setRank] = useState<Rank>();
   const [page, setPage] = useState<number>(1);
   const pageInfo = {
@@ -32,7 +28,7 @@ export default function RankList({
     setPage,
   };
 
-  const makePath = () => {
+  const makePath = (): string => {
     const modeQuery = (targetMode?: string) =>
       targetMode !== 'normal' ? 'ranks/single' : 'vip';
     const seasonQuery = toggleMode === 'rank' ? `&season=${season}` : '';
@@ -41,43 +37,15 @@ export default function RankList({
       : `/pingpong/${modeQuery(toggleMode)}?page=${page}${seasonQuery}`;
   };
 
-  useEffect(() => {
-    async function waitRankList() {
-      await getRankDataHandler();
-    }
-
-    waitRankList().then(() => {
-      if (isScroll) {
-        window.scrollTo({
-          top: ((myRank[toggleMode] - 1) % 20) * 45,
-          behavior: 'smooth',
-        });
-        setIsScroll(false);
-      }
-    });
-  }, [page, isScroll]);
-
-  useEffect(() => {
-    page !== 1
-      ? ((pageInfo.currentPage = 1), setPage(1))
-      : getRankDataHandler();
-  }, [toggleMode, season]);
-
-  useEffect(() => {
-    if (isScroll) {
-      setPage(Math.ceil(myRank[toggleMode] / 20));
-    }
-  }, [isScroll]);
-
-  const getRankDataHandler = async () => {
-    try {
-      const res = await instance.get(`${makePath()}`);
-      setRank(res?.data);
-      setMyRank((prev) => ({ ...prev, [toggleMode]: res?.data.myRank }));
-    } catch (e) {
-      setError('DK01');
-    }
-  };
+  useRankList([
+    makePath(),
+    toggleMode,
+    season,
+    setRank,
+    page,
+    setPage,
+    pageInfo,
+  ]);
 
   if (isMain) return <RankListMain rank={rank} />;
